\name{cstar-package}
\alias{cstar-package}
\alias{cstar-package}
\docType{package}
\title{
Functions to examine the substantive robustness of regression-based results.
}
\description{
Researchers care not only about the statistical significance of results, but also their substantive significance. This package implements a set of tools that allow researchers to use Bayesian decision-theoretic techniques set out in Esarey and Danneman (2014) to determine the substantive robustness of regression-based results.  Specifically, given a set of regression results, a researcher can determine the minimum substantively meaningful relationship supported by his or her evidence under a variety levels of risk aversion.
}
\details{
\tabular{ll}{
Package: \tab cstar\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-13\cr
License: \tab GPL-2\cr
}

}
\author{
Justin Esarey
Maintainer: Justin Esarey <justin@justinesarey.com>
}

\references{
Esarey and Danneman (2014).  A Quantitative Method for Substantive Robustness Assessment.  Political Science Research and Methods.
}

\keyword{ cstar }

